Fixed bug in getting the PSF width.

+++ instrument_modules/src/instruments.cpp
@@ -74,8 +74,9 @@ void Instrument::common_constructor(Json::Value noise_pars){
   
   int pix_x  = specs["psf"]["pix_x"].asInt();
   int pix_y  = specs["psf"]["pix_y"].asInt();
-  int width  = specs["psf"]["width"].asDouble();
-  int height = specs["psf"]["height"].asDouble();
+  double width  = specs["psf"]["width"].asDouble();
+  double height = specs["psf"]["height"].asDouble();
+  
   this->original_psf = new RectGrid(pix_x,pix_y,0,width,0,height,full_path+"psf.fits");
 
   this->noise = FactoryNoiseModel::getInstance()->createNoiseModel(noise_pars,this);
